+++ tests/util/sandbox/test_sandbox_files.py
@@ -1,35 +1,53 @@
 from pathlib import Path
 
+import pytest
+from test_helpers.utils import skip_if_no_docker
+
 from inspect_ai import Task, eval
-from inspect_ai.agent._human.agent import human_cli
-from inspect_ai.agent._react import react
 from inspect_ai.dataset import Sample
 from inspect_ai.dataset._dataset import MemoryDataset
-from inspect_ai.tool._tools._execute import bash
+from inspect_ai.solver._solver import Solver, solver
+from inspect_ai.util._sandbox.context import sandbox
 
-FILES_DIR = Path(__name__).parent / "test_sandbox_files"
-
-
-# @skip_if_no_docker
-# @pytest.mark.slow
-def test_docker_sandbox_files() -> None:
-    dataset = MemoryDataset(
-        [
-            Sample(
-                input="Is there a directory named 'dir1' in the working directory? What are its contents?",
-                files={"dir1": (FILES_DIR / "dir1").as_posix()},
-            )
-        ]
-    )
-
-    task = Task(
-        dataset=dataset,
-        solver=react(tools=[bash()]),
-        sandbox="docker",
-    )
-
-    eval(task, model="openai/gpt-4o", solver=human_cli())[0]
+FILES_DIR = Path(__file__).parent / "test_sandbox_files"
 
 
-if __name__ == "__main__":
-    test_docker_sandbox_files()
+@skip_if_no_docker
+@pytest.mark.slow
+def test_docker_sandbox_files_dir() -> None:
+    def check_sandbox_files_dir(files_dir: Path) -> None:
+        files = {files_dir.name: files_dir.as_posix()}
+        dataset = MemoryDataset([Sample(input="Here are some files", files=files)])
+
+        task = Task(
+            dataset=dataset,
+            solver=[verify_files(files_dir)],
+            sandbox="docker",
+        )
+
+        log = eval(task, model="mockllm/model")[0]
+        assert log.status == "success"
+
+    check_sandbox_files_dir(FILES_DIR / "dir1")
+    check_sandbox_files_dir(FILES_DIR / "dir1" / "dirA")
+
+
+@solver
+def verify_files(files_dir: Path) -> Solver:
+    # list of relative file paths
+    relative_paths = [
+        f"{files_dir.name}/{str(path.relative_to(files_dir))}"
+        for path in files_dir.rglob("*")
+        if path.is_file()
+    ]
+
+    async def solve(state, generate):
+        for path in relative_paths:
+            try:
+                await sandbox().read_file(path)
+            except FileNotFoundError:
+                assert False
+
+        return state
+
+    return solve